+++ api/views/device/controlPanel.ejs
@@ -10,10 +10,15 @@
             padding: 0 2rem;
             gap: 1rem;
          }
+         .container {
+            margin: auto;
+            text-align: center;
+         }
 
          .button-container {
             margin-top: 20px;
             display: flex;
+            justify-content: center;
             gap: 1rem;
          }
 
@@ -78,7 +83,6 @@
             <h1>Control Panel <b><%= id %></b></h1>
          </div>
          <% const binary = device.state.toString(2).padStart(3, "0"); %>
-         <hr />
          <div class="button-container">
             <label class="toggle">
                <span class="label">Light</span>
